Feat: menyelesaikan revisi

+++ src/app/components/akutansi/listrekening/(table)/deleteButton.tsx
@@ -2,6 +2,7 @@
 
 import { deleteData } from "@/app/(public)/(main)/akutansi/listrekening/action"
 import useStore from "@/app/(public)/(main)/akutansi/listrekening/store"
+import { passwordInputAlert } from "@/utils/alert/swal"
 import { useState } from "react"
 import Swal from "sweetalert2"
 
@@ -12,10 +13,14 @@ const DeleteButton = ({ id }: {
     const reset = useStore(state => state.reset)
     const handleClick = async () => {
         try{
-            setIsLoading(true)
-            await deleteData([id])
-            setIsLoading(false)
-            reset()
+            const confirmPassword = await passwordInputAlert()
+
+            if(confirmPassword){
+                setIsLoading(true)
+                await deleteData([id])
+                setIsLoading(false)
+                reset()
+            }
         }catch{
             Swal.fire({
                 icon: 'error',
